Stream dashboard analytics sections
